Refactor side_menu, add menu_item
