+++ src/renderer/components/side_menu/menu_item.js
@@ -1,0 +1,25 @@
+var React = require('react');
+
+var MenuItem = React.createClass({
+  handleClick: function(e) {
+    e.preventDefault();
+    this.props.swapView(this.props.name);
+  },
+
+  render: function() {
+    var className = 'main-nav-item ';
+    if (this.props.active) {
+      className += 'active ';
+    }
+    var titleClassName = this.props.name + ' main-nav-title';
+
+    return (
+      <li className={className} onClick={this.handleClick}>
+        <img data-src={this.props.icon} className='iconic iconic-md main-nav-icon' />
+        <p className={titleClassName}>{this.props.name}</p>
+      </li>
+    );
+  }
+});
+
+module.exports = MenuItem;
